Update requirements of numpy  (#2006)
### Changes

Updated to support version of numpy to 1.24,
- Updated requirements `numpy>=1.19.1, <1.24` to `numpy>=1.19.1, <1.25` 
- Remove using deprecated aliases
https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations.
- Remove tests `mobilefacedet-v1-mxnet` model from mxnet, that does not
support newer versions of numpy.


### Related tickets

115936

+++ nncf/torch/utils.py
@@ -132,7 +132,7 @@ def fp32_accum_wrapper(func):
     def wrapper(tensor_to_sum, ret_tensor):
         half = tensor_to_sum.dtype == np.float16
         if half:
-            tensor_to_sum = tensor_to_sum.astype(np.float)
+            tensor_to_sum = tensor_to_sum.astype(np.float32)
         retval = func(tensor_to_sum, ret_tensor)
         if half:
             retval = retval.astype(np.float16)
